style: remove InformationalHeader from settings screen, adjust styling and responsivity across different components
